Extract shared inner classes to top-level
